config_tools: remove UUID and use vmname to identify a vm
1. remove CONFIG_KATA_VM, CONFIG_MAX_KATA_VM_NUM and KATA_VM
related code.
2. remove KATA_VM in scenario xml files.
3. remove KATA vm type in UI.
4. remove UUID DB.
5. remove uuid in the launch script generation logic.
6. use vmname to identify a vm. (eg: <name>ACRN_SOS_VM</name>)
add the check logic to ensure vm name is a must-set item, not duplicated,
and the string length is [1-15] and the whitespace is not allowed.
7. call default_populator.py to expand the default value in the
scenario XML file to fix the issue that don't export xml successfully in UI
because that the unexpanded xml does not conform to schema.

Signed-off-by: Kunhui-Li <kunhuix.li@intel.com>